fix(clippy)+test(memory): C+E 并行完成
C: clippy warnings 206→69（B 修了 137 个：unused imports/closures/casts/clones）
E: tests/memory_injection_ci.sh — Memory save/search/list/forget CI 脚本

全量测试：655 passed

+++ src/agent/compact.rs
@@ -72,7 +72,7 @@ where
     for attempt in 0..=config.max_retries {
         if attempt > 0 {
             let delay = std::time::Duration::from_millis(
-                500u64 * 2u64.pow((attempt - 1) as u32).min(6),
+                500u64 * 2u64.pow((attempt - 1)).min(6),
             );
             tokio::time::sleep(delay).await;
         }
